Improve error handling when interrupting task's input upload

+++ inductiva/api/methods.py
@@ -4,6 +4,7 @@ The relevant function for usage outside of this file is invoke_async_api().
 Check the demos directory for examples on how it is used.
 """
 import os
+import sys
 import time
 import tqdm
 import tqdm.utils
@@ -162,7 +163,8 @@ def upload_input(api_instance: TasksApi, task_id, original_params,
         logging.info("")
 
     finally:
-        os.remove(input_zip_path)
+        if input_zip_path:
+            os.remove(input_zip_path)
 
 
 def block_until_finish(api_instance: TasksApi, task_id: str) -> str:
@@ -191,7 +193,7 @@ def kill_task(api_instance: TasksApi, task_id: str):
    """
     logging.debug("Sending kill task request ...")
     api_instance.kill_task(path_params={"task_id": task_id},)
-    logging.info("Task terminated.")
+    logging.info("Task with ID %s was terminated.", task_id)
 
 
 def get_task_status(api_instance: TasksApi, task_id: str) -> TaskStatus:
@@ -245,7 +247,9 @@ def block_until_status_is(api_instance: TasksApi,
 
 
 @contextmanager
-def blocking_task_context(api_instance: TasksApi, task_id):
+def blocking_task_context(api_instance: TasksApi,
+                          task_id: str,
+                          action_str: str = "action"):
     """Context to handle execution of a blocking task.
 
     The context handles exceptions and the SIGINT signal, issuing a request
@@ -268,10 +272,10 @@ def blocking_task_context(api_instance: TasksApi, task_id):
         logging.info("Caught exception: terminating blocking task...")
         kill_task(api_instance, task_id)
         raise err
-    except KeyboardInterrupt as err:
-        logging.info("Caught SIGINT: terminating blocking task...")
+    except KeyboardInterrupt:
+        logging.info("Caught SIGINT: %s interrupted by user.", action_str)
         kill_task(api_instance, task_id)
-        raise err
+        sys.exit(1)
     finally:
         # Reset original SIGINT handler
         signal.signal(signal.SIGINT, original_sig)
@@ -360,7 +364,7 @@ def submit_task(api_instance,
 
     if task_submitted_info["status"] == "pending-input":
         # Use the blocking task context
-        with blocking_task_context(api_instance, task_id):
+        with blocking_task_context(api_instance, task_id, "input upload"):
             upload_input(
                 api_instance=api_instance,
                 original_params=params,
